Updated dev branch, AI branch, and My own branch with Hugo's + Main's changes
